+++ nym-vpn-apple/Home/Sources/26/AppFeature/AppFeatureViewModel.swift
@@ -380,8 +380,7 @@ import GRPCManager
 
     public func requestWelcome() {
         path = NavigationPath()
-        pendingDrawerContent = .welcome
-        drawerContent = .welcome
+        stagePreauthDrawer(.welcome)
     }
 
     public func requestInactiveSubscriptionPurchase() {