+++ nym-vpn-apple/NymVPN/NymVPNApp.swift
@@ -57,10 +57,7 @@ struct NymVPNApp: App {
     @ObservedObject private var gatewayManager = GatewayManager.shared
     @ObservedObject private var impactGenerator = ImpactGenerator.shared
     @ObservedObject private var purchasesManager = PurchasesManager()
-    @State private var deeplinkManager = DeeplinkManager(
-        credentialsManager: CredentialsManager.shared,
-        connectionManager: ConnectionManager.shared
-    )
+    @State private var deeplinkManager = DeeplinkManager(credentialsManager: CredentialsManager.shared)
 
     @State private var appFeatureViewModel = AppFeatureViewModel(
         appSettings: .shared,