+++ nym-vpn-apple/Services/Package.swift
@@ -139,9 +139,7 @@ let package = Package(
             name: "DeeplinkManager",
             dependencies: [
                 "CredentialsManager",
-                "ConnectionManager",
-                .product(name: "Constants", package: "ServicesMutual"),
-                .product(name: "ConnectionTypes", package: "ServicesMutual")
+                .product(name: "Constants", package: "ServicesMutual")
             ],
             path: "Sources/Services/DeeplinkManager"
         ),