+++ nym-vpn-apple/Services/Sources/Services/ConnectionManager/ConnectionManager+macOS.swift
@@ -1,11 +1,8 @@
 #if os(macOS)
 import Foundation
 import ConnectionTypes
-import Constants
 import NotificationMessages
 import TunnelMixnet
-import TunnelStatus
-import WidgetKit
 
 extension ConnectionManager {
     @MainActor func connect() async throws {
@@ -44,8 +41,6 @@ extension ConnectionManager {
 
 extension ConnectionManager {
     func setupGRPCManagerObservers() {
-        updateWidgetState(for: currentTunnelStatus)
-
         grpcManager.$isServing.sink { [weak self] isConnectedToDaemon in
             guard isConnectedToDaemon else { return }
 
@@ -61,7 +56,6 @@ extension ConnectionManager {
                 self?.currentTunnelStatus = status
                 self?.scheduleNotificationIfNeeded()
                 self?.updateTimeConnected()
-                self?.updateWidgetState(for: status)
             }
         }
         .store(in: &cancellables)
@@ -130,25 +124,6 @@ private extension ConnectionManager {
         Task {
             await NotificationMessages.scheduleDisconnectNotification()
         }
-    }
-}
-
-// MARK: - Widget -
-extension ConnectionManager {
-    func updateWidgetState(for status: TunnelStatus) {
-        let defaults = UserDefaults(suiteName: Constants.groupID.rawValue)
-        if status == .connected {
-            if let code = connectionStorage.entryGateway.countryCode {
-                let name = Locale.current.localizedString(forRegionCode: code) ?? code
-                defaults?.set(name, forKey: "macos_widgetEntryLocation")
-            }
-            if let code = connectionStorage.exitRouter.countryCode {
-                let name = Locale.current.localizedString(forRegionCode: code) ?? code
-                defaults?.set(name, forKey: "macos_widgetExitLocation")
-            }
-        }
-        defaults?.set(status.rawValue, forKey: "macos_widgetTunnelStatus")
-        WidgetCenter.shared.reloadAllTimelines()
     }
 }
 #endif