+++ nym-vpn-apple/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2026.12.0] - TBD
 
+### Removed
+
+- Widget support for both iOS & macOS
+
 ### Fixed
 
 - Fix switching between modes to always land on the last selected mode (https://github.com/nymtech/nym-vpn-client/pull/6023)